requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
google-search-results==2.4.0
idna==3.3
numpy==1.19.5
pandas==1.1.5
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
tqdm==4.62.3
urllib3==1.26.7
